+++ src/interval_tree/find/entry.rs
@@ -2,13 +2,15 @@ use std::ops::RangeInclusive;
 
 #[derive(Debug)]
 pub struct Entry<'a, K: Clone + Ord, V> {
-    #[deprecated(note = "use `entry.key()` instead")]
-    pub key: &'a RangeInclusive<K>,
-    #[deprecated(note = "use `entry.get()` instead")]
-    pub value: &'a V,
+    key: &'a RangeInclusive<K>,
+    value: &'a V,
 }
 
 impl<'a, K: Clone + Ord, V> Entry<'a, K, V> {
+    pub(crate) fn new(key: &'a RangeInclusive<K>, value: &'a V) -> Self {
+        Self { key, value }
+    }
+
     /// Returns a reference to the key in the entry.
     pub fn key(&self) -> &RangeInclusive<K> {
         #[allow(deprecated)]