+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   * Intersections include the endpoints, e.g., [0, 1] and [3, 5] now intersect
     [1, 3].
 
+  * `Entry` fields `key` and `value` are now private. Use `entry.key()` and
+    `entry.get()`, respectively, instead.
+
 ## 0.2.0 - 2020-05-08
 
 ### Changed